+++ diskover/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Diskover v2 Community Edition Change Log
 
+# [2.0-rc.3] - 2021-12-
+### fixed
+- if an unhandled error occurred, diskover would not exit without keyboard interupt
+- exception when using alt scanners
+### added
+- indices now tokenize camel case in file names and paths
+- optional function name "init" used by alt scanners to set up connections to api, get env vars, etc.
+- optional function name "close" used by alt scanners to close dbs, etc.
+- --threads cli option, overrides maxthreads config setting
+### changed
+- maxthreads diskover config settings now default to auto set based on number of cpus when leaving config setting blank, see default/sample config file
+
+
 # [2.0-rc.2] - 2021-12-01
 ### fixed
 ### added